+++ qmmp/src/plugins/Input/mad/decoder_mad.h
@@ -43,7 +43,7 @@ private:
     void deinit();
     bool findHeader();
     bool findXingHeader(struct mad_bitptr, unsigned int);
-    uint findID3v2(uchar *data, uint size);
+    uint findID3v2(uchar *data, ulong size);
     bool inited, user_stop, done, m_finish, derror, eof, useeq;
     qint64 totalTime, seekTime;
     int channels;
